+++ README.rst
@@ -5,14 +5,14 @@ Back translation
    :target: https://github.com/mutalyzer/backtranslate/graphs/commit-activity
 .. image:: https://travis-ci.org/mutalyzer/backtranslate.svg?branch=master
    :target: https://travis-ci.org/mutalyzer/backtranslate
-.. image:: https://readthedocs.org/projects/simplerpc/badge/?version=latest
-   :target: https://backtranslate.readthedocs.io/en/latest
+.. image:: https://readthedocs.org/projects/mutalyzer-backtranslate/badge/?version=latest
+   :target: https://mutalyzer-backtranslate.readthedocs.io/en/latest
 .. image:: https://img.shields.io/github/release-date/mutalyzer/backtranslate.svg
    :target: https://github.com/mutalyzer/backtranslate/releases
 .. image:: https://img.shields.io/github/release/mutalyzer/backtranslate.svg
    :target: https://github.com/mutalyzer/backtranslate/releases
-.. image:: https://img.shields.io/pypi/v/backtranslate.svg
-   :target: https://pypi.org/project/backtranslate/
+.. image:: https://img.shields.io/pypi/v/mutalyzer-backtranslate.svg
+   :target: https://pypi.org/project/mutalyzer-backtranslate/
 .. image:: https://img.shields.io/github/languages/code-size/mutalyzer/backtranslate.svg
    :target: https://github.com/mutalyzer/backtranslate
 .. image:: https://img.shields.io/github/languages/count/mutalyzer/backtranslate.svg
@@ -30,4 +30,4 @@ nucleotides.
 Please see ReadTheDocs_ for the latest documentation.
 
 
-.. _ReadTheDocs: https://backtranslate.readthedocs.io
+.. _ReadTheDocs: https://mutalyzer-backtranslate.readthedocs.io
