+++ docs/api.rst
@@ -1,0 +1,15 @@
+API documentation
+=================
+
+Main library
+------------
+
+.. automodule:: mutalyzer_backtranslate.backtranslate
+   :members:
+
+
+Utils
+-----
+
+.. automodule:: mutalyzer_backtranslate.util
+   :members:
