+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = backtranslate
-version = 0.1.2
+name = mutalyzer_backtranslate
+version = 0.1.3
 description = Functions for reverse translation.
 long_description = file: README.rst
 author = LUMC, Jeroen F.J. Laros
@@ -28,4 +28,4 @@ tests_require =
 
 [options.entry_points]
 console_scripts =
-    backtranslate = backtranslate.cli:main
+    backtranslate = mutalyzer_backtranslate.cli:main
